requirements.txt
# fastapi
# uvicorn[standard]
# pydantic
# datasets
# pandas
# numpy
# scikit-learn
# imbalanced-learn
# category_encoders
# flaml
# xgboost
# joblib
# faker
# matplotlib
# python-dotenv
# requests





fastapi
uvicorn[standard]
pydantic
datasets
pandas
numpy
scikit-learn
imbalanced-learn
flaml
xgboost
lightgbm
catboost
joblib
faker
matplotlib
python-dotenv
requests
beautifulsoup4
transformers
huggingface-hub

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Additional ML libraries
optuna
kaggle

# Web server for Azure
gunicorn
